Add separate tests for parameter resolution, and simplify build_parameters() tests.
